% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_data.R
\docType{data}
\name{sal_data}
\alias{sal_data}
\title{Length-mass data}
\format{
Data frame
}
\usage{
data(sal_data)
}
\description{
Length-mass data collected from red-backed salamanders in Ohio
}
\keyword{datasets}
