% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smi.R
\name{smi}
\alias{smi}
\title{Calculate standardized mass index (SMI)}
\usage{
smi(mass, length, L0 = NULL, resid = FALSE)
}
\arguments{
\item{mass}{A vector of mass values}

\item{length}{A vector of length values}

\item{L0}{A constant value used in standardization. If not specified, the mean of `length` is used}

\item{resid}{(Default = FALSE) If `TRUE`, the residuals of ordinary least squares regression also be returned}
}
\value{
A data frame containing the calculated SMI, and optionally, OLS residuals
}
\description{
Calculates SMI using raw length and mass data
}
\details{
This function calculates SMI following Peig and Green. Additionally, the function can return residuals from OLS regression, as an alternative measure / index of body condition. Provide untransformed length and mass values. They will be log-transformed during the calculation of SMI. It is also not necessary to remove observations with missing data values.

Peig, J., and A. J. Green. 2009. New perspectives for estimating body condition from mass/length data: The scaled mass index as an alternative method. Oikos 118:1883–1891.
}
\examples{
## Not run:
data("sal_data")

smi_results1 <- smi(mass = sal_data$Mass,
                    length = sal_data$SVL)

smi_results2 <- smi(mass = sal_data$Mass,
                    length = sal_data$SVL,
                    L0 = mean(sal_data$SVL, na.rm = TRUE))

smi_results3 <- smi(mass = sal_data$Mass,
                    length = sal_data$SVL,
                    L0 = mean(sal_data$SVL, na.rm = TRUE),
                    resid = TRUE)

plot(smi_results3$smi ~ smi_results3$resid)

## End (Not run)

}
\author{
Bill Peterman <Peterman.73@osu.edu>
}
